feat(@theme-ui/mdx): add TypeScript support
