+++ packages/mdx/tsconfig.json
@@ -1,0 +1,4 @@
+{
+    "extends": "../core/tsconfig.json",
+    "include": ["src/**/*.ts", "src/**/*.tsx", "src/**/*.d.ts"]
+}
